% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggarnet_mn.R
\name{ggarnet_mn}
\alias{ggarnet_mn}
\title{Title}
\usage{
ggarnet_mn()
}
\arguments{
\item{type}{= database name of the desired plot
to see plots a available check data file}
}
\value{
a ggplot plot with the selected polygon
}
\description{
Title
}
\examples{
ggarnet_mn()
}
