% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggarnet_tourmaline.R
\name{ggarnet_tourmaline}
\alias{ggarnet_tourmaline}
\title{ggarnet_tourmaline}
\usage{
ggarnet_tourmaline()
}
\arguments{
\item{type}{= database name of the desired plot
to see plots a available check data file}
}
\value{
a ggplot plot with the selected polygon
}
\description{
ggarnet_tourmaline
}
\examples{
ggarnet_tourmaline()
}
